#5: Updated docker images and added workflow to deploy json server
