+++ .env.development
@@ -1,1 +1,1 @@
-VITE_BACKEND_API_URL=http://localhost:3001
+VITE_BACKEND_API_URL=http://localhost:9000
