+++ .env.production
@@ -1,1 +1,2 @@
+# URL of the backend API on azure
 VITE_BACKEND_API_URL=/api
