+++ Dockerfile
@@ -1,6 +1,6 @@
 # ==== BUILD =====
 # Use NodeJS v18.18.2 with Alpine for a slim image
-FROM node:18.18.2-alpine3.18 as build
+FROM node:18-alpine as build
 LABEL authors="theGreyPilgrim42"
 
 # Set the working directory
@@ -20,7 +20,12 @@ RUN npm run build
 
 # ==== RUN =====
 # Use stable Nginx version with Alpine for a slim image
-FROM nginx:stable-alpine3.17-slim
+FROM nginx:stable-alpine-slim
+
+# Delete the default nginx.conf
+RUN rm /etc/nginx/conf.d/default.conf  # <= This line solved my issue
+
+COPY src/config/nginx.conf /etc/nginx/conf.d
 
 # Copy the build files to the nginx html directory
 COPY --from=build /react-docker/dist /usr/share/nginx/html
