+++ api/Dockerfile
@@ -1,21 +1,21 @@
 # ==== BUILD =====
 # Use NodeJS v18.18.2 with Alpine for a slim image
-FROM node:18.18.2-alpine3.18 as build
-LABEL authors="theGreyPilgrim42"
+FROM node:18-alpine as build
+LABEL authors="SmashGrade"
 
-EXPOSE 3000
+EXPOSE 9000
 
 # Set the working directory
-WORKDIR /react-docker/
+WORKDIR /json-server/
 
 # Copy app files
 COPY . .
 
-# Set the working directory to api
-WORKDIR /react-docker/api
-
 # Install dependencies from package-lock.json
 RUN npm install-clean
 
+# Build typescript to javascript
+RUN npm run build
+
 # Start the development json server
-CMD ["npm", "run", "start:docker"]
+CMD ["npm", "run", "start"]
