+++ api/package.json
@@ -1,15 +1,19 @@
 {
-  "name": "react-app-with-json-server",
+  "name": "json-server",
   "version": "1.0.0",
   "engines": {
     "node": ">=18.0.0"
   },
   "scripts": {
-    "start": "npm i && node server.js",
-    "start:docker": "node serverDocker.js"
+    "build": "tsc",
+    "start": "node server.js"
   },
   "dependencies": {
     "cors": "^2.8.5",
     "json-server": "^0.17.4"
+  },
+  "devDependencies": {
+    "@types/json-server": "^0.14.7",
+    "typescript": "^5.2.2"
   }
 }
