+++ api/server.ts
@@ -2,12 +2,17 @@ import * as jsonServer from 'json-server';
 
 const server = jsonServer.create();
 const router = jsonServer.router('db.json');
-const middlewares = jsonServer.defaults({ static: '.' });
+const middlewares = jsonServer.defaults();
 
-const port = process.env.PORT ?? 3000;
+const port = process.env.PORT ?? 9000;
 
 server.use(middlewares);
-server.use('/api', router);
+server.use(router);
+server.use((req, res, next) => {
+    res.setHeader('Access-Control-Allow-Origin', '*');
+    res.setHeader('Accept', '*');
+    next();
+});
 server.listen(port, () => {
-    console.info(`JSON Server is running on: ${port}/api`);
+    console.info(`JSON Server API is running on port: ${port}`);
 });
