+++ package.json
@@ -9,22 +9,23 @@
     "scripts": {
         "dev": "vite",
         "build": "tsc && vite build",
+        "build-local": "tsc && vite build --mode development",
+        "build:azure": "npm run build && tsc api/server.ts --outDir dist",
         "lint": "eslint --ext ts,tsx,css --report-unused-disable-directives --max-warnings 0 src",
         "lint-report": "eslint --ext ts,tsx,css --format json-relative --output-file reports/eslint.json src",
         "format": "prettier --write \"*.{js,jsx,ts,tsx,json,css,scss,md}\"",
         "check-typescript": "tsc --noEmit && tsc api/server.ts --noEmit",
-        "preview": "vite preview",
+        "preview": "concurrently \"npm:json-db\" \"vite preview\"",
         "prepare": "husky install",
-        "json-db": "json-server --watch db.json --port 3001",
+        "json-db": "cd api && npm run start",
         "dev-db": "concurrently \"npm:json-db\" \"npm:dev\"",
-        "start": "json-server db.json --static ./dist",
+        "start": "json-server api/db.json --static ./dist",
         "extract-de": "formatjs extract \"src/**/*.{ts,tsx}\" --ignore=\"**/*.d.ts\" --id-interpolation-pattern \"[sha512: contenthash:base64:6]\" --out-file \"src/i18n/lang/de.json\"",
         "extract-en": "formatjs extract \"src/**/*.{ts,tsx}\" --ignore=\"**/*.d.ts\" --id-interpolation-pattern \"[sha512: contenthash:base64:6]\" --out-file \"src/i18n/lang/en.json\"",
         "extract-fr": "formatjs extract \"src/**/*.{ts,tsx}\" --ignore=\"**/*.d.ts\" --id-interpolation-pattern \"[sha512: contenthash:base64:6]\" --out-file \"src/i18n/lang/fr.json\"",
         "compile-de": "formatjs compile \"src/i18n/lang/de.json\" --out-file \"src/i18n/compiled-lang/de.json\"",
         "compile-en": "formatjs compile \"src/i18n/lang/en.json\" --out-file \"src/i18n/compiled-lang/en.json\"",
-        "compile-fr": "formatjs compile \"src/i18n/lang/fr.json\" --out-file \"src/i18n/compiled-lang/fr.json\"",
-        "build:azure": "npm run build && tsc api/server.ts --outDir dist"
+        "compile-fr": "formatjs compile \"src/i18n/lang/fr.json\" --out-file \"src/i18n/compiled-lang/fr.json\""
     },
     "lint-staged": {
         "*.{js,jsx,ts,tsx,json,css,scss,md}": [
