+++ src/config/nginx.conf
@@ -1,0 +1,17 @@
+server {
+    listen 80;
+    location / {
+      root   /usr/share/nginx/html;
+      index  index.html index.htm;
+      try_files $uri $uri/ /index.html;
+    }
+    error_page 404 /index.html;
+    location = / {
+      root /usr/share/nginx/html;
+      internal;
+    }
+    error_page   500 502 503 504  /50x.html;
+    location = /50x.html {
+      root   /usr/share/nginx/html;
+    }
+  }
